requirements.txt
fastapi
uvicorn[standard]
mistralai[realtime]
python-dotenv
python-multipart
